ML Translation Evaluation/requirements.txt
setuptools==69.0.0
sacrebleu
nltk
pandas
torch
torchmetrics==0.10.3
unbabel-comet
numpy
